add one form both for creat and edit note
