+++ src/App.css
@@ -27,14 +27,11 @@ body {
   overflow-y: auto;
 }
 
-h1 {
-  margin-bottom: 20px;
-}
-
 table {
   width: 100%;
   border-collapse: collapse;
-  margin-top: 20px;
+  margin-top: 10px;
+  margin-bottom: 10px;
 }
 
 table th,
@@ -62,7 +59,7 @@ table td:nth-child(5) {
   width: 220px;
 }
 
-table td:nth-child(5)>div {
+table td:nth-child(5) > div {
   display: flex;
   justify-content: center;
 }
@@ -92,8 +89,9 @@ button[type="button"]:hover {
   background-color: #999;
 }
 
+h1,
 h2 {
-  margin-top: 40px;
+  margin: 10px 0px 0px 0px;
 }
 
 form {
@@ -114,7 +112,6 @@ select {
   border-radius: 4px;
   margin-bottom: 10px;
 }
-
 
 button[type="submit"],
 button[type="button"] {
@@ -139,11 +136,29 @@ button[type="button"] {
   border-radius: 8px;
   box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.2);
   width: 500px; /* Adjust the width as needed */
-  height: 250px; /* Adjust the height as needed */
+}
+
+.modal-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
 }
 
 .modal h2 {
-  margin-top: 0;
+  margin: 0;
+}
+
+.modal-close {
+  cursor: pointer;
+  font-size: 24px;
+  color: #aaa;
+  border: none;
+  background-color: transparent;
+}
+
+.modal-close:hover {
+  color: #555;
 }
 
 .modal-footer {
