+++ src/containers/NotesPage.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { RootState } from '../types/types';
 import NoteTable from '../components/NoteTable';
-import NoteForm from '../components/NoteForm';
 import SummaryTable from '../components/SummaryTable';
 import { updateSummary } from '../actions/noteActions';
 
@@ -19,7 +18,6 @@ const NotesPage: React.FC = () => {
     <div className="app-container">
       <h1>Notes App</h1>
       <div className="note-page-content">
-        <NoteForm />
         <NoteTable notes={notes} />
         <SummaryTable summaryData={summaryData} />
       </div>
